% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{beagle2pl}
\alias{beagle2pl}
\title{Format a beagle file as phred-scaled likelihoods.}
\usage{
beagle2pl(file, lines = NULL, ind_order = NULL, ind_names = NULL)
}
\arguments{
\item{file}{character. File path to the beagle or beagle.gz file containing
the raw data.}

\item{lines}{numeric, default NULL. Optional vector of numbers noting which
lines in the beagle file to keep. This mut be sequential due to
\code{\link[data.table]{fread}}'s restirctions.}

\item{ind_order}{numeric, default NULL. Numeric vector indicating final
sorted order to use for individuals the output matrix (\code{c(2,1,3)}, for
example, would put the second individual in \code{file} first, then the
first individual second, and then the third individual last in the final
matrix, for example). If NULL, the order will be unchanged.}

\item{ind_names}{character, default NULL. Names for individuals in the final
matrix. If \code{ind_order} is provided, this should match the final order
dictated by \code{ind_order}; otherwise, this should match the order
in \code{file}.}
}
\value{
A \eqn{p \times n} matrix, where p is the number of loci and n is
the number of individuals. See details for format.
}
\description{
Takes a beagle formatted input file and converts it to a matrix of phred-
scaled likelihoods, as needed by \code{likeGEA}. See details.
}
\details{
The input file format required by this package has one row for each locus and
one column for each individual. Phred-scaled likelihoods (PL) for the three
possible genotypes at each locus/individual are comma seperated (major
homozygote, heterozygote, minor heterozygote). A higher PL constitutes a
lower likelihood; by convention the minimum PL is subtracted from each
triplet such that the most likely genotype has a PL of 0.

Phred-likelihoods can be calculated according to:
\deqn{PL(\text{genotype}) = -10 \times \log_{10}\left( P(\text{Data} \mid \text{Genotype}) \right)}

Where \eqn{P(\text{Data} \mid \text{Genotype})} is the a genotype likelihood
as produced by something like \code{ANGSD}'s \code{doGlf}.
}
\author{
William Hemstrom

pl <- function("path/to/example/data/once/Nikunj/adds/it.beagle.gz",
lines = 1:2000) # first 2000 lines
}
